Replace styled-components with emotion
Styled components v5 was blowing up every test in the library with
invalid hook errors. It could have something to do with the org
of the style js files and how they are imported. Regardless, there
shouldn't be anything wrong with the existing setup. A switch to
emotion fixed the suite so going with that as I had been thinking
about migrating to that framework anyway.